Make methods in BaseComponent be protected - they should only be used by components/pages.
